absorb: move core functionality from cli to lib
